requirements.txt
celery==5.3.6
Flask==3.0.2
python-dotenv==1.0.1
